Move sedov box and constant functions to separate files
