+++ sedov_blast_wave/sedov_box.hpp
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "../wash_mockapi.hpp"
+
+void apply_pbc(const double h, double& xx, double& yy, double& zz);
+double distance_pbc(const double h, const wash::Particle& p, const wash::Particle& q);
