+++ sedov_blast_wave/sedov_consts.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+#define _USE_MATH_DEFINES
+#include <cmath>
+
+extern const double k;
+void init_wh();
+double lookup_wh(const double v);
